Create 0000-mobile-nft-creation-fee.md
This HIP requires the creation and holding of a non-fungible token (NFT) for devices that do not pay an onboarding fee on the MOBILE subDAO. Whichever wallet holds the NFT will receive applicable MOBILE rewards for that device. NFTs for active devices will vary in range, and the NFT Creation burn fee is noted in the detailed explanation section below. 

As this HIP creates moves radios on chain in the form of NFTs, this HIP also changes the rewardable entity of MOBILE from MOBILE Hotspots to radios.